Doodle Math
Doodle Math is a math quiz game powered by Scratch, a game that will benefit children and primary school students in mathematics and mind development, and solve the issue of children's addiction to brain-rotting content and unhelpful content that is common in our world, especially Generation Alpha, by simply turning the addiction of smart devices into playing a math game that can develop childrens' minds and improving the ability of quick-decisions and acting fast, besides learning mathematics.

The project has aspects that are beginner-friendly and exciting for children and others who are familiar with smart devices. by using an easy-to-understand user interface for kids and adults, colorful characters with reactivity while playing, and various difficulties with game modes for accessibility. 

I hope you have an enjoyable experience playing this game.